+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Zero is now falsy in if/else conditions as it is in Handlebars.js
 * Lex number literals that start with 0
 * Hashes now work with #each blocks
+* Zero is now truthy in unless conditions as it is in Handlebars.js
 
 ## [1.0.0] - 2022-01-19